requirements.txt
pySerial
pybase64
untangle